requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#

autoflake==1.4
    # via -r requirements.in
black==22.3.0
    # via -r requirements.in
click==8.1.3
    # via black
flake8==4.0.1
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
pillow==7.2.0
    # via -r requirements.in
platformdirs==2.5.2
    # via black
pycodestyle==2.8.0
    # via flake8
pydantic==1.9.0
    # via -r requirements.in
pyflakes==2.4.0
    # via
    #   autoflake
    #   flake8
tomli==2.0.1
    # via black
typing-extensions==4.2.0
    # via
    #   black
    #   pydantic
